[Cli] Reorganize factory classes in Interaction namespace.
