+++ src/Valkyrja/Cli/Interaction/Argument/Factory/ArgumentFactory.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Cli\Interaction\Factory;
+namespace Valkyrja\Cli\Interaction\Argument\Factory;
 
 use Valkyrja\Cli\Interaction\Argument\Argument;
