+++ src/Valkyrja/Cli/Interaction/Output/Factory/Contract/OutputFactoryContract.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Cli\Interaction\Factory\Contract;
+namespace Valkyrja\Cli\Interaction\Output\Factory\Contract;
 
 use Valkyrja\Cli\Interaction\Enum\ExitCode;
 use Valkyrja\Cli\Interaction\Message\Contract\MessageContract;
