+++ src/Valkyrja/Cli/Routing/Controller/Controller.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace Valkyrja\Cli\Routing\Controller;
 
-use Valkyrja\Cli\Interaction\Factory\Contract\OutputFactoryContract;
 use Valkyrja\Cli\Interaction\Input\Contract\InputContract;
+use Valkyrja\Cli\Interaction\Output\Factory\Contract\OutputFactoryContract;
 
 abstract class Controller
 {
