+++ tests/Unit/Cli/Interaction/Argument/Factory/ArugmentFactoryTest.php
@@ -11,9 +11,9 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Tests\Unit\Cli\Interaction\Factory;
+namespace Valkyrja\Tests\Unit\Cli\Interaction\Argument\Factory;
 
-use Valkyrja\Cli\Interaction\Factory\ArgumentFactory;
+use Valkyrja\Cli\Interaction\Argument\Factory\ArgumentFactory;
 use Valkyrja\Tests\Unit\Abstract\TestCase;
 
 /**
